requirements.txt
langchain
langchain-community
requests
python-dotenv
openai
beautifulsoup4
